Rename test files to follow source code directoy structure
