+++ test/unit/db/repos/transactions/dapp.js
@@ -14,27 +14,27 @@
 
 'use strict';
 
-const DBSandbox = require('../../../common/db_sandbox').DBSandbox;
-const transactionsFixtures = require('../../../fixtures').transactions;
-const seeder = require('../../../common/db_seed');
-const transactionTypes = require('../../../../helpers/transaction_types');
+const DBSandbox = require('../../../../common/db_sandbox').DBSandbox;
+const transactionsFixtures = require('../../../../fixtures/index').transactions;
+const seeder = require('../../../../common/db_seed');
+const transactionTypes = require('../../../../../helpers/transaction_types');
 
 const numSeedRecords = 5;
 
 let db;
 let dbSandbox;
-let multiSignaturesRepo;
+let dappRepo;
 
 describe('db', () => {
 	before(done => {
 		dbSandbox = new DBSandbox(
 			__testContext.config.db,
-			'lisk_test_db_transactions_multisignature'
+			'lisk_test_db_transactions_dapp'
 		);
 
 		dbSandbox.create((err, __db) => {
 			db = __db;
-			multiSignaturesRepo = db['transactions.multisignature'];
+			dappRepo = db['transactions.dapp'];
 			done(err);
 		});
 	});
@@ -60,58 +60,54 @@ describe('db', () => {
 	});
 
 	it('should initialize db.blocks repo', () => {
-		return expect(multiSignaturesRepo).to.be.not.null;
+		return expect(dappRepo).to.be.not.null;
 	});
 
-	describe('MultiSigTransactionsRepo', () => {
+	describe('DappsTransactionsRepo', () => {
 		describe('constructor()', () => {
 			it('should assign param and data members properly', () => {
-				expect(multiSignaturesRepo.db).to.be.eql(db);
-				expect(multiSignaturesRepo.pgp).to.be.eql(db.$config.pgp);
-				expect(multiSignaturesRepo.dbTable).to.be.eql('multisignatures');
-				expect(multiSignaturesRepo.dbFields).to.be.eql([
-					'min',
-					'lifetime',
-					'keysgroup',
+				expect(dappRepo.db).to.be.eql(db);
+				expect(dappRepo.pgp).to.be.eql(db.$config.pgp);
+				expect(dappRepo.dbTable).to.be.eql('dapps');
+				expect(dappRepo.dbFields).to.be.eql([
+					'type',
+					'name',
+					'description',
+					'tags',
+					'link',
+					'category',
+					'icon',
 					'transactionId',
 				]);
 
-				expect(multiSignaturesRepo.cs).to.be.an('object');
-				expect(multiSignaturesRepo.cs).to.not.empty;
-				return expect(multiSignaturesRepo.cs).to.have.all.keys('insert');
+				expect(dappRepo.cs).to.be.an('object');
+				expect(dappRepo.cs).to.not.empty;
+				return expect(dappRepo.cs).to.have.all.keys('insert');
 			});
 		});
 
 		describe('save', () => {
-			it('should insert entry into "delegates" table for type 4 transactions', function*() {
+			it('should insert entry into "dapps" table for type 5 transactions', function*() {
 				const block = seeder.getLastBlock();
 				const transactions = [];
 				for (let i = 0; i < numSeedRecords; i++) {
 					transactions.push(
 						transactionsFixtures.Transaction({
 							blockId: block.id,
-							type: transactionTypes.MULTI,
+							type: transactionTypes.DAPP,
 						})
 					);
 				}
 				yield db.transactions.save(transactions);
 
-				const result = yield db.query('SELECT * FROM multisignatures');
+				const result = yield db.query('SELECT * FROM dapps');
 
 				expect(result).to.not.empty;
 				expect(result).to.have.lengthOf(numSeedRecords);
-				expect(result.map(r => r.min)).to.be.eql(
-					transactions.map(t => t.asset.multisignature.min)
-				);
-				expect(result.map(r => r.lifetime)).to.be.eql(
-					transactions.map(t => t.asset.multisignature.lifetime)
-				);
-				expect(result.map(r => r.keysgroup)).to.be.eql(
-					transactions.map(t => t.asset.multisignature.keysgroup.join())
-				);
-				return expect(result.map(r => r.transactionId)).to.be.eql(
+				expect(result.map(r => r.transactionId)).to.be.eql(
 					transactions.map(t => t.id)
 				);
+				return expect(result).to.be.eql(transactions.map(t => t.asset.dapp));
 			});
 		});
 	});
