+++ test/unit/db/repos/transactions/in_transfer.js
@@ -14,27 +14,27 @@
 
 'use strict';
 
-const DBSandbox = require('../../../common/db_sandbox').DBSandbox;
-const transactionsFixtures = require('../../../fixtures').transactions;
-const seeder = require('../../../common/db_seed');
-const transactionTypes = require('../../../../helpers/transaction_types');
+const DBSandbox = require('../../../../common/db_sandbox').DBSandbox;
+const transactionsFixtures = require('../../../../fixtures/index').transactions;
+const seeder = require('../../../../common/db_seed');
+const transactionTypes = require('../../../../../helpers/transaction_types');
 
 const numSeedRecords = 5;
 
 let db;
 let dbSandbox;
-let outTransferRepo;
+let inTransferRepo;
 
 describe('db', () => {
 	before(done => {
 		dbSandbox = new DBSandbox(
 			__testContext.config.db,
-			'lisk_test_db_transactions_out_transfer'
+			'lisk_test_db_transactions_in_transfer'
 		);
 
 		dbSandbox.create((err, __db) => {
 			db = __db;
-			outTransferRepo = db['transactions.outTransfer'];
+			inTransferRepo = db['transactions.inTransfer'];
 			done(err);
 		});
 	});
@@ -60,53 +60,46 @@ describe('db', () => {
 	});
 
 	it('should initialize db.blocks repo', () => {
-		return expect(outTransferRepo).to.be.not.null;
+		return expect(inTransferRepo).to.be.not.null;
 	});
 
-	describe('OutTransferTransactionsRepo', () => {
+	describe('InTransferTransactionsRepo', () => {
 		describe('constructor()', () => {
 			it('should assign param and data members properly', () => {
-				expect(outTransferRepo.db).to.be.eql(db);
-				expect(outTransferRepo.pgp).to.be.eql(db.$config.pgp);
-				expect(outTransferRepo.dbTable).to.be.eql('outtransfer');
-				expect(outTransferRepo.dbFields).to.be.eql([
-					'dappId',
-					'outTransactionId',
-					'transactionId',
-				]);
+				expect(inTransferRepo.db).to.be.eql(db);
+				expect(inTransferRepo.pgp).to.be.eql(db.$config.pgp);
+				expect(inTransferRepo.dbTable).to.be.eql('intransfer');
+				expect(inTransferRepo.dbFields).to.be.eql(['dappId', 'transactionId']);
 
-				expect(outTransferRepo.cs).to.be.an('object');
-				expect(outTransferRepo.cs).to.not.empty;
-				return expect(outTransferRepo.cs).to.have.all.keys('insert');
+				expect(inTransferRepo.cs).to.be.an('object');
+				expect(inTransferRepo.cs).to.not.empty;
+				return expect(inTransferRepo.cs).to.have.all.keys('insert');
 			});
 		});
 
 		describe('save', () => {
-			it('should insert entry into "delegates" table for type 7 transactions', function*() {
+			it('should insert entry into "delegates" table for type 6 transactions', function*() {
 				const block = seeder.getLastBlock();
 				const transactions = [];
 				for (let i = 0; i < numSeedRecords; i++) {
 					transactions.push(
 						transactionsFixtures.Transaction({
 							blockId: block.id,
-							type: transactionTypes.OUT_TRANSFER,
+							type: transactionTypes.IN_TRANSFER,
 						})
 					);
 				}
 				yield db.transactions.save(transactions);
 
-				const result = yield db.query('SELECT * FROM outtransfer');
+				const result = yield db.query('SELECT * FROM intransfer');
 
 				expect(result).to.not.empty;
 				expect(result).to.have.lengthOf(numSeedRecords);
 				expect(result.map(r => r.transactionId)).to.be.eql(
 					transactions.map(t => t.id)
 				);
-				expect(result.map(r => r.dappId)).to.be.eql(
-					transactions.map(t => t.asset.outTransfer.dappId)
-				);
-				return expect(result.map(r => r.outTransactionId)).to.be.eql(
-					transactions.map(t => t.asset.outTransfer.transactionId)
+				return expect(result.map(r => r.dappId)).to.be.eql(
+					transactions.map(t => t.asset.inTransfer.dappId)
 				);
 			});
 		});
