+++ test/unit/db/repos/transactions/multisignature.js
@@ -14,27 +14,27 @@
 
 'use strict';
 
-const DBSandbox = require('../../../common/db_sandbox').DBSandbox;
-const transactionsFixtures = require('../../../fixtures').transactions;
-const seeder = require('../../../common/db_seed');
-const transactionTypes = require('../../../../helpers/transaction_types');
+const DBSandbox = require('../../../../common/db_sandbox').DBSandbox;
+const transactionsFixtures = require('../../../../fixtures/index').transactions;
+const seeder = require('../../../../common/db_seed');
+const transactionTypes = require('../../../../../helpers/transaction_types');
 
 const numSeedRecords = 5;
 
 let db;
 let dbSandbox;
-let transferRepo;
+let multiSignaturesRepo;
 
 describe('db', () => {
 	before(done => {
 		dbSandbox = new DBSandbox(
 			__testContext.config.db,
-			'lisk_test_db_transactions_transfer'
+			'lisk_test_db_transactions_multisignature'
 		);
 
 		dbSandbox.create((err, __db) => {
 			db = __db;
-			transferRepo = db['transactions.transfer'];
+			multiSignaturesRepo = db['transactions.multisignature'];
 			done(err);
 		});
 	});
@@ -60,43 +60,54 @@ describe('db', () => {
 	});
 
 	it('should initialize db.blocks repo', () => {
-		return expect(transferRepo).to.be.not.null;
+		return expect(multiSignaturesRepo).to.be.not.null;
 	});
 
-	describe('TransferTransactionsRepo', () => {
+	describe('MultiSigTransactionsRepo', () => {
 		describe('constructor()', () => {
 			it('should assign param and data members properly', () => {
-				expect(transferRepo.db).to.be.eql(db);
-				expect(transferRepo.pgp).to.be.eql(db.$config.pgp);
-				expect(transferRepo.dbTable).to.be.eql('transfer');
-				expect(transferRepo.dbFields).to.be.eql(['data', 'transactionId']);
+				expect(multiSignaturesRepo.db).to.be.eql(db);
+				expect(multiSignaturesRepo.pgp).to.be.eql(db.$config.pgp);
+				expect(multiSignaturesRepo.dbTable).to.be.eql('multisignatures');
+				expect(multiSignaturesRepo.dbFields).to.be.eql([
+					'min',
+					'lifetime',
+					'keysgroup',
+					'transactionId',
+				]);
 
-				expect(transferRepo.cs).to.be.an('object');
-				expect(transferRepo.cs).to.not.empty;
-				return expect(transferRepo.cs).to.have.all.keys('insert');
+				expect(multiSignaturesRepo.cs).to.be.an('object');
+				expect(multiSignaturesRepo.cs).to.not.empty;
+				return expect(multiSignaturesRepo.cs).to.have.all.keys('insert');
 			});
 		});
 
 		describe('save', () => {
-			it('should insert entry into "delegates" table for type 0 transactions', function*() {
+			it('should insert entry into "delegates" table for type 4 transactions', function*() {
 				const block = seeder.getLastBlock();
 				const transactions = [];
 				for (let i = 0; i < numSeedRecords; i++) {
 					transactions.push(
 						transactionsFixtures.Transaction({
 							blockId: block.id,
-							type: transactionTypes.TRANSFER,
+							type: transactionTypes.MULTI,
 						})
 					);
 				}
 				yield db.transactions.save(transactions);
 
-				const result = yield db.query('SELECT * FROM transfer');
+				const result = yield db.query('SELECT * FROM multisignatures');
 
 				expect(result).to.not.empty;
 				expect(result).to.have.lengthOf(numSeedRecords);
-				expect(result.map(r => Buffer.from(r.data).toString())).to.be.eql(
-					transactions.map(t => t.asset.data)
+				expect(result.map(r => r.min)).to.be.eql(
+					transactions.map(t => t.asset.multisignature.min)
+				);
+				expect(result.map(r => r.lifetime)).to.be.eql(
+					transactions.map(t => t.asset.multisignature.lifetime)
+				);
+				expect(result.map(r => r.keysgroup)).to.be.eql(
+					transactions.map(t => t.asset.multisignature.keysgroup.join())
 				);
 				return expect(result.map(r => r.transactionId)).to.be.eql(
 					transactions.map(t => t.id)
